+++ Pulp/Common/Logging/FilterPolicies/LoggerNoFilterPolicy.inl
@@ -18,7 +18,7 @@ inline void LoggerNoFilterPolicy::RegisterVars(void)
 }
 
 inline bool LoggerNoFilterPolicy::Filter(const char* type, const SourceInfo& sourceInfo, 
-	const char* channel, size_t verbosity, const char* format, va_list args)
+	const char* channel, int verbosity, const char* format, va_list args)
 {
 	X_UNUSED(type);
 	X_UNUSED(sourceInfo);
